Fine-tune the sample size for CMRR.py (#136)
* Fine-tune sample size for CMRR.py

So here's the idea: the amount of time it takes to run CMRR for small values of days_to_simulate is very small. Which means that we can increase the sample size without having to worry about CMRR taking too long to annoy the user.
Instead of writing a whole bunch of else-if statements, I made a function that calculates the "best" sample size. Specifically, it calculates what sample size would make CMRR run roughly as long as it owuld with sample_size=4 and days_to_simulate=365.
sample_size=4 and days_to_simulate=365 is used as the default, and then if days_to_simulate is lower, sample size is increased to make the total run time approximately constant.

* Update fsrs_simulator.py

* Update fsrs_simulator.py

* Change the coefficients.py

* Fixed a problem where the output was 3 instead of 4.py

* Update fsrs_simulator.py

I don't know how I managed to delete "best" twice today

* Update fsrs_simulator.py

* set loss_aversion to 1 when plotting workload

* Update ylim.py

Don't mind me, I'm just trying things

* Update fsrs_simulator.py

* min -> minimum.py

* Decrease ylim.py

* Update fsrs_simulator.py

Ensure that nothing is plotted above the box with the graph

* bump version

---------

Co-authored-by: Jarrett Ye <jarrett.ye@outlook.com>